# Conflicts:
#	.env.example
#	core-service/src/test/java/com/moodit/core_service/controller/CourseControllerTest.java
#	frontend/src/components/UserMenu/UserMenu.tsx
#	frontend/src/context/CurrentUserProvider.tsx
#	frontend/src/index.css
#	frontend/src/pages/Dashboard/Dashboard.tsx

+++ core-service/src/main/java/com/moodit/core_service/dto/QuestionResultDTO.java
@@ -14,8 +14,8 @@ import java.util.List;
 @AllArgsConstructor
 public class QuestionResultDTO {
     private Integer questionId;
-    private Integer earned;
-    private Integer max;
+    private Double earned;
+    private Double max;
     /** Choix : ids corrects / ids choisis. */
     private List<Integer> correctAnswerIds;
     private List<Integer> selectedAnswerIds;
